feat: added error handling in case blogPosts array is null on both home and blog pages
