+++ bootcamp-milestone-2/src/app/globals.css
@@ -337,3 +337,7 @@ h3.entry-title-end {
   margin-bottom: 30px;
   text-align: center;
 }
+
+.noBlogs {
+  padding-bottom: 100px;
+}
